requirements.txt
# TabBacklog v1 - Python Dependencies

# Web Frameworks
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6

# Database
asyncpg==0.29.0
psycopg[binary]==3.1.18
sqlalchemy==2.0.25

# HTTP Client
httpx==0.26.0
requests==2.31.0

# HTML/XML Parsing
beautifulsoup4==4.12.3
lxml==5.1.0
html5lib==1.1

# Video/Content Extraction
yt-dlp==2024.1.0

# LLM & DSPy
dspy-ai==2.4.0
openai==1.10.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Vector Search
pgvector==0.2.4

# CLI
click==8.1.7
rich==13.7.0
python-dotenv==1.0.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Utilities
tenacity==8.2.3  # Retry logic
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx[testing]==0.26.0

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0

# Markdown Export
markdown==3.5.2
